move test-only BatchScope functionality to BatchTestScope
