+++ ledger/txtail_test.go
@@ -154,6 +154,7 @@ func (t *txTailTestLedger) initialize(ts *testing.T, protoVersion protocol.Conse
 	t.protoVersion = protoVersion
 
 	err := t.trackerDBs.Batch(func(transactionCtx context.Context, tx store.BatchScope) (err error) {
+		testTx := tx.(store.BatchTestScope)
 		arw, err := tx.MakeAccountsWriter()
 		if err != nil {
 			return err
@@ -161,7 +162,7 @@ func (t *txTailTestLedger) initialize(ts *testing.T, protoVersion protocol.Conse
 
 		accts := ledgertesting.RandomAccounts(20, true)
 		proto := config.Consensus[protoVersion]
-		newDB := tx.AccountsInitTest(ts, accts, protoVersion)
+		newDB := testTx.AccountsInitTest(ts, accts, protoVersion)
 		require.True(ts, newDB)
 
 		roundData := make([][]byte, 0, proto.MaxTxnLife)
